split serializers into folder structure
