--- fitness/serializers.py
@@ -1,18 +1,0 @@
-from rest_framework import serializers
-from .models import Account, Activity
-
-
-class AccountSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Account
-        fields = [
-            "id",
-            "username",
-            "first_name",
-        ]
-
-
-class ActivitySerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Activity
-        fields = '__all__'
