+++ fitness/serializers/activity.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from ..models import Activity
+
+class ActivitySerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Activity
+        fields = '__all__'
